requirements.txt
pbr>=5.8.0 # Apache-2.0
fastapi==0.111.0 # MIT
pydantic>=1.9.0,<2.0.0
uvicorn==0.30.1 # BSD License (3 clause)
gunicorn>=20.1.0 # MIT
python-jose<=3.3.0 # MIT
databases>=0.9.0 # BSD License (3 clause)
aiomysql>=0.2.0 # MIT
aiosqlite>=0.20.0 # MIT
loguru<=0.5.3 # MIT
PyYAML>=5.4.1 # MIT
immutables>=0.16 # Apache-2.0
alembic>=1.7.5 # MIT
httpx>=0.16.1 # BSD License (3 clause)
SQLAlchemy>=1.3.24 # MIT
PyMySQL>=0.9.3 # MIT
dnspython>=2.1.0 # ISC
click>=7.1.2 # BSD License (3 clause)
jinja2>=2.11.3 # BSD License (3 clause)
h11>=0.11 # MIT
MarkupSafe>=2.0.1 # BSD License (3 clause)
python-multipart>=0.0.5 # Apache-2.0
python-keystoneclient>=3.21.0 # Apache-2.0
python-cinderclient>=8.1.0 # Apache-2.0
python-glanceclient>=2.17.1 # Apache-2.0
python-neutronclient>=6.14.1 # Apache-2.0
python-novaclient>=15.1.1 # Apache-2.0
keystoneauth1>=3.17.4 # Apache-2.0
oslo.policy>=2.3.4 # Apache-2.0
types-six>=1.16.0 # Apache-2.0
